Fix a bug with regex used for parsing yarn.lock.
The previous version of the regex incorrectly looked for exact package names preceded by any character which resulted in false matches (e.g. `react-dom@18.2.0` matched `@types/react-dom@18.2.0`). This resulted in a crash when assuming that npmjs.com API would contain the date for a non-existent version of the package. Now the regex requires that the package name be preceded by a space, a quote, or be placed at the very beginning of the line.