Fix an issue that produced invalid Bundler spec definitions.
When specifying a path to another Ruby project while already being in a Ruby project, gems and gem groups were incorrectly loading from the project from where the script was run instead of using the path of the project to be audited, which was caused by a missing Bundler reset statement. This mismatch caused gems to be assigned to incorrect groups or not be assigned to any groups.