Remove the need to do `bundle install` before Ruby auditing.
Ruby gems are now resolved remotely and instead of locally, which eliminates the annoying requirements to have the gems be installed locally and potentially having to install a new Ruby version and struggle with the installation of gems with native extension.

Run package-audit from anywhere on any Ruby project, and all the gems will be resolved remotely.