Remove Bundler warning about an old bundler gem version.
Previously, the following warning would be generated if the installed bundler gem version is older than that within Gemfile.lock:

Evaluating packages and their dependencies... -Warning: the running version of Bundler (2.1.4) is older than the version that created the lockfile (2.3.8). We suggest you to upgrade to the version that created the lockfile by running `gem install bundler:2.3.8`.